jobque: marker args serialize signed; palette modulus from length (review)
traceSave emitted marker args with %u while the API takes a signed int —
a negative arg round-tripped as a huge unsigned value. Emit %d with a
bit-preserving int32 cast, and pin it with a negative-arg case in the
round-trip test. resolve_color's modulus now uses length(DEFAULT_PALETTE)
instead of a hardcoded 12.

+++ src/misc/job_que.cpp
@@ -318,11 +318,12 @@ namespace das {
             for ( uint32_t i = 0; i != L.n; ++i ) {
                 const auto & e = L.ev[i];
                 if ( e.tag == uint32_t(TraceTag::Marker) ) {
-                    // perfetto instant event; name = the registered marker kind
+                    // perfetto instant event; name = the registered marker kind. arg is
+                    // signed in the API (jobque_trace_marker) — emit %d so it round-trips
                     fprintf(f, ",\n{\"ph\":\"i\",\"s\":\"g\",\"pid\":0,\"tid\":%d,\"ts\":%.3f,\"name\":",
                         int(lane), double(e.t0 - tmin) / 1000.0);
                     fputsJsonString(f, e.stage < uint32_t(markerNames.size()) ? markerNames[e.stage].c_str() : "marker");
-                    fprintf(f, ",\"args\":{\"arg\":%u}}", e.arg);
+                    fprintf(f, ",\"args\":{\"arg\":%d}}", int32_t(e.arg));
                     continue;
                 }
                 fprintf(f, ",\n{\"ph\":\"X\",\"pid\":0,\"tid\":%d,\"ts\":%.3f,\"dur\":%.3f,\"name\":\"%s\","
